Make the background color of .notice less intense because DANG
Bright colors on a dark, intense background have intense, unpleasant
visual side-effects; muting the background makes for a more pleasant and
legible experience.

I set the alpha of that purple much lower, which fixed the problem. Then
I set the RGB values a little bluer and less purple, which looked a bit
better against the red and white.